Convert views to use mustache templates
This change largely affects the way templates are rendered server-side:
it renders mustache templates instead of my weird string replace
function.

Another change snuck in here is bundling the client-side javascript
using rollup. There's a nice bit of middleware called rollup-express
that allows for a route to be defined as a rollup target. It serves the
file from memory instead of creating a file and recompiles fresh on each
request.